requirements.txt
requests-html
python-docx
lxml_html_clean